Detect recovery from failed notifications
This attempts to detect if we can recover from failed notifications by
checking if we go from failed notification to a non-failed notification
(for the same keys).

In addition this adds the persistentId of a failed notification to our
list of handled push notifications so that we don't try to keep
processing the same failed notification.